Annotate captchas with their labels

Given one or more Captchas, this function prompts the user to solve them mannually to train a model. Annotated captchas are saved at path with their labels in the filename separated by an underscore.

Usage

captcha_annotate(files, labels = NULL, path = NULL, rm_old = FALSE)

Arguments

files

Either an object of class captcha or a character vector with the paths to captcha files

labels

Either NULL (for interactive classification) or a character vector with labels for the Captchas. See details.

path

Where to save the annotated captcha files. If NULL, saves the files in the same folder the unanswered counterparts.

rm_old

Whether or not to delete unanswered captchas after copying and renaming them.

Value

A vector with the paths of the modified files.

Details

The labels= parameter can handle situations where one knows the Captcha label. For example, a workflow that uses an oracle might provide the label automatically. When the label doesn't exist, the captcha_annotate() function opens the prompt for classification and shows the image using plot().