Net captcha

This is a torch module with This is a simple CNN with three convolutional layers and two dense layers. It works well with the default dimensions 32x192 from the captcha_transform_image() function. It also uses batch normalization in the forward method. This function can be used either to fit a Captcha model using the luz workflow suggested in vignette("advanced") or as a base code to develop custom models.

Usage

net_captcha(
  input_dim,
  output_ndigits,
  output_vocab_size,
  vocab,
  transform,
  dropout = c(0.25, 0.25),
  dense_units = 400
)

Arguments

input_dim

(integer, integer): image input dimensions.

output_ndigits

number of tokens for each Captcha.

output_vocab_size

number of unique token values.

vocab

token labels

transform

input transform function (for prediction purposes)

dropout

(float, float) AlexNet dropout values.

dense_units

Number of dense units

Value

object of classes CAPTCHA-CNN and nn_module. It works as a predictive function and as the input to a luz fitting workflow.

Examples

# raw image
captcha_file <- system.file(
  "examples/captcha/tjmg.jpeg",
  package = "captcha"
)

# initializes model. The output_ndigits, output_vocab_size and vocab
# parameters are compatible to the TJMG Captcha.
model <- net_captcha(
  input_dim = c(32, 192),
  output_ndigits = 5,
  output_vocab_size = 10,
  vocab = 0:9,
  transform = captcha_transform_image,
  dropout = c(.25, .25),
  dense_units = 400
)
model
#> An `nn_module` containing 1,204,296 parameters.
#> 
#> ── Modules ─────────────────────────────────────────────────────────────────────
#> • batchnorm0: <nn_batch_norm2d> #6 parameters
#> • conv1: <nn_conv2d> #896 parameters
#> • batchnorm1: <nn_batch_norm2d> #64 parameters
#> • conv2: <nn_conv2d> #18,496 parameters
#> • batchnorm2: <nn_batch_norm2d> #128 parameters
#> • conv3: <nn_conv2d> #36,928 parameters
#> • batchnorm3: <nn_batch_norm2d> #128 parameters
#> • dropout1: <nn_dropout> #0 parameters
#> • dropout2: <nn_dropout> #0 parameters
#> • fc1: <nn_linear> #1,126,800 parameters
#> • batchnorm_dense: <nn_batch_norm1d> #800 parameters
#> • fc2: <nn_linear> #20,050 parameters

transformed <- model$transform(captcha_file)

# tensor of size 1 x output_ndigits x output_vocab_size
prediction <- model(transformed)
dim(prediction)
#> [1]  1  5 10

# get the predicted labels
# the model is awful, because it is not fitted yet.
indices <- as.numeric(torch::torch_argmax(prediction, 3))
label <- paste(model$vocab[indices], collapse = "")